requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
branca==0.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
django-bootstrap4==24.4
django-crispy-forms==2.3
django-leaflet==0.31.0
django-phonenumber-field==8.0.0
folium==0.19.4
geographiclib==2.0
geopy==2.4.1
idna==3.10
Jinja2==3.1.5
jwcrypto==1.5.6
jwt==1.3.1
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.2.2
phonenumbers==8.13.54
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-jwt==4.1.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xyzservices==2025.1.0
